feat: 커뮤니티 상단 헤더 컴포넌트로 변경, 게시판 생성 로직 구현, 게시물 구현중
