+++ app/(community)/tube/[communityid]/page.tsx
@@ -106,7 +106,6 @@ function Tube({ params } : { params : { communityid : string }}) {
 
         </div>
       </>
-     
   )
 }
