+++ app/creation/posting/[boardId]/page.tsx
@@ -1,12 +1,14 @@
 import React from 'react'
 import Posting from '@/components/Posting'
 
-function PostingPage() {
+function PostingPage({params: {boardId}} : {params: {boardId: string}}) {
+
   return (
     <div>
-      <Posting  />
+      <Posting boardId={boardId}/>
     </div>
   )
-}
 
+
+}
 export default PostingPage
