+++ styles/globals.css
@@ -61,7 +61,7 @@
   grid-column: -1 / 13; /* 그리드의 처음부터 끝까지 확장 */
 }
 
-/* 커뮤니티 페이지 상단부 */
+/* 커뮤니티 페이지 상단부
 .community-header {
   font-family: 'Arial', sans-serif;
   color: #333;
@@ -72,4 +72,4 @@
   justify-content: center;
   align-items: center;
   flex-direction: column;
-}
+} */
